Refactor | restructure the data models
