+++ src/query_executor/pp_executor.py
@@ -1,4 +1,4 @@
-from src.models import FrameBatch
+from src.models.storage.batch import FrameBatch
 
 
 class PPExecutor:
