+++ test/loaders/test_video_loader.py
@@ -5,7 +5,10 @@ import cv2
 import numpy as np
 
 from src.loaders.video_loader import SimpleVideoLoader
-from src.models import VideoMetaInfo, VideoFormat, Frame, ColorSpace, FrameInfo
+from src.models.catalog.frame_info import FrameInfo
+from src.models.catalog.properties import ColorSpace, VideoFormat
+from src.models.catalog.video_info import VideoMetaInfo
+from src.models.storage.frame import Frame
 
 NUM_FRAMES = 10
